Now base classes for models are out of fastoad.models
More consistent: these classes have to part of FAST-OAD core, unlike the actual models.
More robust: Having these classes embedded in the models package will lead to import loop if fastoad.models is declared as plugin (which is about to be done).
