+++ src/fastoad/model_base/__init__.py
@@ -1,6 +1,3 @@
-"""
-Package for propulsion modules
-"""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
 #  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
