+++ src/fastoad/models/aerodynamics/aerodynamics_landing.py
@@ -2,7 +2,7 @@
 Aero computation for landing phase
 """
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -17,7 +17,7 @@ Aero computation for landing phase
 import numpy as np
 import openmdao.api as om
 
-from fastoad.models.options import OpenMdaoOptionDispatcherGroup
+from fastoad.model_base.options import OpenMdaoOptionDispatcherGroup
 from fastoad.utils.physics import Atmosphere
 from .components.compute_max_cl_landing import ComputeMaxClLanding
 from .components.high_lift_aero import ComputeDeltaHighLift
