+++ src/fastoad/models/performances/breguet/breguet.py
@@ -1,6 +1,6 @@
 """Implementation of the Breguet Formula."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -17,7 +17,7 @@ from scipy.constants import g
 
 from fastoad.base.flight_point import FlightPoint
 from fastoad.constants import EngineSetting
-from fastoad.models.propulsion import IPropulsion
+from fastoad.model_base.propulsion import IPropulsion
 from fastoad.utils.physics import AtmosphereSI
