+++ src/fastoad/models/performances/breguet/openmdao.py
@@ -1,6 +1,6 @@
 """Simple Breguet-based module for performances."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -17,7 +17,7 @@ import openmdao.api as om
 from scipy.constants import g
 
 from fastoad.constants import EngineSetting
-from fastoad.models.propulsion.fuel_propulsion.base import FuelEngineSet
+from fastoad.model_base.propulsion import FuelEngineSet
 from fastoad.module_management.service_registry import RegisterPropulsion
 from fastoad.utils.physics import AtmosphereSI
 from . import Breguet
