+++ src/fastoad/models/performances/mission/mission_definition/mission_builder.py
@@ -2,7 +2,7 @@
 Mission generator.
 """
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -21,11 +21,11 @@ import openmdao.api as om
 
 from fastoad.base.flight_point import FlightPoint
 from fastoad.constants import EngineSetting
+from fastoad.model_base.propulsion import IPropulsion
 from fastoad.models.performances.mission.base import FlightSequence
 from fastoad.models.performances.mission.base import IFlightPart
 from fastoad.models.performances.mission.polar import Polar
 from fastoad.models.performances.mission.segments.base import FlightSegment
-from fastoad.models.propulsion import IPropulsion
 from .schema import (
     PHASE_DEFINITIONS_TAG,
     ROUTE_DEFINITIONS_TAG,
