+++ src/fastoad/models/performances/mission/openmdao/sizing_mission.py
@@ -2,7 +2,7 @@
 OpenMDAO component for computation of sizing mission.
 """
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -22,7 +22,7 @@ import openmdao.api as om
 from scipy.constants import foot
 
 from fastoad.base.flight_point import FlightPoint
-from fastoad.models.propulsion.fuel_propulsion.base import FuelEngineSet
+from fastoad.model_base.propulsion import FuelEngineSet
 from fastoad.module_management.service_registry import RegisterPropulsion
 from . import resources
 from .mission_wrapper import MissionWrapper
