+++ src/fastoad/models/performances/mission/segments/base.py
@@ -1,6 +1,6 @@
 """Base classes for simulating flight segments."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -24,7 +24,7 @@ from scipy.optimize import root_scalar
 
 from fastoad.base.flight_point import FlightPoint
 from fastoad.constants import EngineSetting
-from fastoad.models.propulsion import IPropulsion
+from fastoad.model_base.propulsion import IPropulsion
 from fastoad.utils.physics import AtmosphereSI
 from ..base import IFlightPart
 from ..exceptions import FastFlightSegmentIncompleteFlightPoint
