+++ src/fastoad/models/performances/mission/segments/tests/test_segments.py
@@ -1,5 +1,5 @@
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -18,8 +18,8 @@ from numpy.testing import assert_allclose
 from scipy.constants import foot
 
 from fastoad.base.flight_point import FlightPoint
+from fastoad.model_base.propulsion import AbstractFuelPropulsion, FuelEngineSet
 from fastoad.models.performances.mission.segments.hold import HoldSegment
-from fastoad.models.propulsion.fuel_propulsion.base import AbstractFuelPropulsion, FuelEngineSet
 from ..altitude_change import AltitudeChangeSegment
 from ..cruise import OptimalCruiseSegment, CruiseSegment, ClimbAndCruiseSegment
 from ..speed_change import SpeedChangeSegment
