+++ src/fastoad/models/propulsion/fuel_propulsion/rubber_engine/openmdao.py
@@ -1,6 +1,6 @@
 """OpenMDAO wrapping of RubberEngine."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -15,7 +15,11 @@
 import numpy as np
 from openmdao.core.component import Component
 
-from fastoad.models.propulsion import IOMPropulsionWrapper, IPropulsion, BaseOMPropulsionComponent
+from fastoad.model_base.propulsion import (
+    IPropulsion,
+    IOMPropulsionWrapper,
+    BaseOMPropulsionComponent,
+)
 from fastoad.openmdao.validity_checker import ValidityDomainChecker
 from .rubber_engine import RubberEngine
