+++ src/fastoad/models/propulsion/fuel_propulsion/rubber_engine/rubber_engine.py
@@ -1,6 +1,6 @@
 """Parametric turbofan engine."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -22,7 +22,7 @@ import pandas as pd
 from fastoad.base.flight_point import FlightPoint
 from fastoad.constants import EngineSetting
 from fastoad.exceptions import FastUnknownEngineSettingError
-from fastoad.models.propulsion.fuel_propulsion.base import AbstractFuelPropulsion
+from fastoad.model_base.propulsion import AbstractFuelPropulsion
 from fastoad.utils.physics import Atmosphere
 from .constants import (
     ALPHA,
