+++ src/fastoad/models/propulsion/fuel_propulsion/tests/test_propulsion.py
@@ -1,5 +1,5 @@
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -18,7 +18,7 @@ import pandas as pd
 from numpy.testing import assert_allclose
 
 from fastoad.base.flight_point import FlightPoint
-from ..base import AbstractFuelPropulsion, FuelEngineSet
+from fastoad.model_base.propulsion import AbstractFuelPropulsion, FuelEngineSet
 
 
 class DummyEngine(AbstractFuelPropulsion):
