+++ src/fastoad/models/weight/mass_breakdown/mass_breakdown.py
@@ -1,6 +1,6 @@
 """Main components for mass breakdown."""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -14,7 +14,7 @@
 
 import openmdao.api as om
 
-from fastoad.models.options import PAYLOAD_FROM_NPAX
+from fastoad.model_base.options import PAYLOAD_FROM_NPAX
 from .a_airframe import (
     WingWeight,
     FuselageWeight,
